This adds support for multiple patch fragments
The JSON config file now must contain an array of patch fragments.
The pipeline stages that used to assume a single patch now have been
modified to iterate the same behavior by folding over the set of patch
fragments in the KB.  Docs and examples have been updated.  There is a
new example, resources/simple-multi, that shows off the new
functionality.

Some of the pipeline phases could use unit tests to check what happens
if the number of patches is 0 or if it changes unexpectedly.