Add loadw to parser for short-term testing of Cody's work
1) This is a temp fix because soon we'll move to Ivan's new primus lisp stuff

2) I expected to get some type errors because I'm mixing bit widths but have
fixed the core theory ghost type to be a specific type variable.  Either I
misunderstand what the type parameter is doing or I did something wrong here.
But it seems to work, and we're going to throw it away next week anyway, so
I'm not going to look to hard at it.