Fixes up handling of register names for preassignment
We now have the `Substituter` add a special prefix to variables which
are known to hold registers (which we would like to preassign as well).
The use of this extends beyond the substitution pass, into `Core_c` as
well as `Bir_passes`.

This also seems to have eliminated the need for the `__Vibes_tmp` prefix
to avoid name clashes between temporaries and registers.