Don't adjust exits when the patch is in a trivial form
We may try to compile a patch that consists of a single `blk term` with
no `jmp term`s. In this case, this (entry block) is also the implicit
exit block, so no new blocks need to be inserted by the `adjust_exits`
pass. This allows us to benefit from more peephole optimizations where
possible.