Fixes semantics of higher variables (#202)
* Remove use of `addw` from the selector

It turns out that with `.syntax unified` the assembler can figure out
whether to use the wide or narrow encoding in Thumb mode.

* Adds the `Naming` submodule to the `Substituter`

* Separates the BIR helpers to a new module

* Don't adjust exits when the patch is in a trivial form

We may try to compile a patch that consists of a single `blk term` with
no `jmp term`s. In this case, this (entry block) is also the implicit
exit block, so no new blocks need to be inserted by the `adjust_exits`
pass. This allows us to benefit from more peephole optimizations where
possible.

* Record spilling information

We're recording the higher variables that were spilled, as well as the
`def term`s that were preserving and/or restoring the variable in
question (that was living in a caller-save register).

* Fixes issue cycles in the scheduler model

We were getting solutions where two or more operations in a block would
have the same issue cycle, which caused dependency clobbering in certain
situations. This constrains the model to pick different issue cycles for
each operation

* Refactors and fixes the semantics of higher variables

* Warn on unused variables in the substituter

* Fix examples in README

Co-authored-by: bmourad01 <bmourad@draper.com>