Adds branch relaxing, splits blocks with conditional jumps
1. We need branch relaxing if the patch needs to perform a conditional
branch to some far away region of code. In this case, the assembler will
fail because the branch is too far away. We break it apart into a
conditional branch to a local block, followed by an unconditional branch
to the desired location if the condition is met.

2. The higher vars initialization may clobber the flags if it occurs at
the same time as a `cmp x, y; bcc loc`. The scheduler is unaware that on
Thumb, many instructions will implicitly set the flags.