Perform naiive branch relaxing (and other fixes to the BIR passes) (#210)
* Collect code addrs for the verifier if `init-mem` is `true`

* Fix failing test

* Update

* Fix use of `init-mem` field

Also pretty-print it with the rest of the WP params

* Only finalize higher vars at the implicit exit

Any explicit exits (blocks that short-circuit the flow of the patch away
from the original flow of the program) are assumed to not need the
higher vars

* Adds the pattern for `ldr rt, [rn, rm, #lsl n]`

Where `n` is a power of two.

It's common to see this with array accesses.

* Adds branch relaxing, splits blocks with conditional jumps

1. We need branch relaxing if the patch needs to perform a conditional
branch to some far away region of code. In this case, the assembler will
fail because the branch is too far away. We break it apart into a
conditional branch to a local block, followed by an unconditional branch
to the desired location if the condition is met.

2. The higher vars initialization may clobber the flags if it occurs at
the same time as a `cmp x, y; bcc loc`. The scheduler is unaware that on
Thumb, many instructions will implicitly set the flags.

* Fix tests and break up the spilling pass

* Cut down minizinc time on `arm-bounds-check`

* Relaxation only applies to Thumb

Co-authored-by: bmourad01 <bmourad@draper.com>