Preserve overwritten instructions if the user tells us to
Some patches require us to only insert new code, which means we will
have to overwrite some instructions that are needed by the program.

The user can tell us this with the `overwrite` field in the patch
info config file. If it's `false` (default `true`) then any
instructions that would be overwritten should be inserted at the
end of the patch code.